requirements.txt
seaborn>=0.13.2
matplotlib>=3.9.2

pillow>=11.0.0
opencv-python>=4.10.0.84
torch==2.5.0+cu124
torchvision>=0.20.0
tqdm>=4.66.5
numpy==1.26.4